negative request num in region load
Test Plan: N/A

Reviewers: cuijianwei

Reviewed By: cuijianwei

Subscribers: infra-dev

Differential Revision: https://phabricator.d.xiaomi.net/D4229

+++ src/main/java/org/apache/hadoop/hbase/regionserver/HRegionServer.java
@@ -1278,7 +1278,7 @@ public class HRegionServer implements HRegionInterface, HBaseRPCErrorHandler,
         storeUncompressedSizeMB,
         storefileSizeMB, memstoreSizeMB, storefileIndexSizeMB, rootIndexSizeKB,
         totalStaticIndexSizeKB, totalStaticBloomSizeKB,
-        (int) r.readRequestsCount.get(), (int) r.writeRequestsCount.get(),
+        r.readRequestsCount.get(), r.writeRequestsCount.get(),
         totalCompactingKVs, currentCompactedKVs, 
         r.getLastFlushSequenceId());
   }
